# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Ratinam Crackers Business Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      security:
        - bearerAuth: []
      parameters:
        - name: invoiceNo
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: string
        - name: agentId
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: financialYear
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create direct invoice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /v1/invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /v1/invoices/{id}/share:
    post:
      operationId: shareInvoice
      tags: [invoices]
      summary: Share invoice via WhatsApp/Email
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice shared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /v1/invoices/gst-report:
    get:
      operationId: getGstReport
      tags: [invoices]
      summary: Get GSTR-1 report data
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: GST report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GstReportResponse"

  # ============================================================
  # PURCHASE ORDERS
  # ============================================================
...
  /v1/packing-jobs:
    get:
      operationId: listPackingJobs
      tags: [reports]
      summary: List packing jobs
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Packing jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackingJobListResponse"
    post:
      operationId: createPackingJob
      tags: [reports]
      summary: Create packing job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackingJobBody"
      responses:
        "201":
          description: Packing job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackingJob"
  /v1/packing-jobs/{id}/complete:
    put:
      operationId: completePackingJob
      tags: [reports]
      summary: Complete packing job
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Packing job completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackingJob"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer

    # AUTH SCHEMAS
...
    EstimateLineItem:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        variantId:
          type: string
        variantSize:
          type: string
        qty:
          type: integer
        resolvedPrice:
          type: number
        resolutionReason:
          type: string
        bulkRateApplied:
          type: boolean
        amount:
          type: number
...
    Invoice:
      type: object
      properties:
        id:
          type: string
        invoiceNo:
          type: string
        customerId:
          type: string
        customerName:
          type: string
        customerGstin:
          type: string
        agentId:
          type: string
        priceListId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/EstimateLineItem"
        subtotal:
          type: number
        discountAmount:
          type: number
        taxableAmount:
          type: number
        cgst:
          type: number
        sgst:
          type: number
        igst:
          type: number
        total:
          type: number
        paymentMode:
          type: string
        channel:
          type: string
        financialYear:
          type: string
        status:
          type: string
        pdfUrl:
          type: string
        createdAt:
          type: string
    InvoiceListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    CreateInvoiceBody:
      type: object
      required: [items, paymentMode]
      properties:
        customerId:
          type: string
        priceListId:
          type: string
        agentId:
          type: string
        locationId:
          type: string
        couponCode:
          type: string
        loyaltyPointsRedeem:
          type: integer
        paymentMode:
          type: string
        channel:
          type: string
        items:
          type: array
          items:
            type: object
            required: [productId, variantId, qty]
            properties:
              productId:
                type: string
              variantId:
                type: string
              qty:
                type: integer
        logisticsDetails:
          type: object
    GstReportResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            b2b:
              type: array
              items:
                type: object
            b2c:
              type: array
              items:
                type: object
            summary:
              type: object

    # PURCHASE ORDER SCHEMAS
...
    PackingJob:
      type: object
      properties:
        id:
          type: string
        jobNumber:
          type: string
        locationId:
          type: string
        rawMaterials:
          type: array
          items:
            type: object
        finishedGoods:
          type: array
          items:
            type: object
        wasteItems:
          type: array
          items:
            type: object
        status:
          type: string
          enum: [pending, in_progress, completed]
        notes:
          type: string
        completedAt:
          type: string
        createdAt:
          type: string
    PackingJobListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/PackingJob"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    CreatePackingJobBody:
      type: object
      required: [locationId, rawMaterials, finishedGoods]
      properties:
        locationId:
          type: string
        rawMaterials:
          type: array
          items:
            type: object
        finishedGoods:
          type: array
          items:
            type: object
        notes:
          type: string

    # DASHBOARD SCHEMAS
...